+++ qontract_development_cli/shell.py
@@ -1,5 +1,4 @@
 import copy
-import functools
 import json
 import logging
 import os
@@ -8,6 +7,7 @@ import sys
 import tempfile
 from multiprocessing import Process
 from pathlib import Path
+from shutil import which
 from typing import Any
 
 from .models import Profile
@@ -19,18 +19,9 @@ from .utils import (
 
 log = logging.getLogger(__name__)
 
-
-@functools.cache
-def _docker_compose_bin() -> list[str]:
-    """Return the docker-compose command. Either 'docker compose' or 'docker-compose'.
-
-    Support the compose plugin and the compose standalone binary.
-    """
-    try:
-        subprocess.run(["docker", "compose", "--help"], check=True, capture_output=True)
-        return ["docker", "compose"]
-    except subprocess.CalledProcessError:
-        return ["docker-compose"]
+_docker_compose_bin = (
+    ["docker-compose"] if which("docker-compose") else ["docker", "compose"]
+)
 
 
 def compose_up(
@@ -40,7 +31,7 @@ def compose_up(
     build: bool = False,
 ) -> None:
     log.info("Starting all containers")
-    compose_cmd = _docker_compose_bin() + ["-f", str(compose_file), "up", "-d"]
+    compose_cmd = _docker_compose_bin + ["-f", str(compose_file), "up", "-d"]
     if force_recreate:
         compose_cmd.append("--force-recreate")
     if remove_orphan:
@@ -52,7 +43,7 @@ def compose_up(
 
 def compose_restart(compose_file: Path, container: str) -> None:
     log.info(f"Restarting {container} container")
-    compose_cmd = _docker_compose_bin() + [
+    compose_cmd = _docker_compose_bin + [
         "-f",
         str(compose_file),
         "restart",
@@ -63,12 +54,12 @@ def compose_restart(compose_file: Path, container: str) -> None:
 
 def compose_down(compose_file: Path) -> None:
     log.info("Stopping all containers")
-    compose_cmd = _docker_compose_bin() + ["-f", str(compose_file), "down"]
+    compose_cmd = _docker_compose_bin + ["-f", str(compose_file), "down"]
     subprocess.run(compose_cmd)
 
 
 def compose_log_tail(compose_file: Path) -> Process:
-    compose_cmd = _docker_compose_bin() + ["-f", str(compose_file), "logs", "--follow"]
+    compose_cmd = _docker_compose_bin + ["-f", str(compose_file), "logs", "--follow"]
     p = EndlessProcess(target=subprocess.run, args=(compose_cmd,))
     p.start()
     return p
@@ -77,7 +68,7 @@ def compose_log_tail(compose_file: Path) -> Process:
 def compose_list_projects() -> list[dict[str, Any]]:
     return json.loads(
         subprocess.run(
-            _docker_compose_bin() + ["ls", "--format", "json"],
+            _docker_compose_bin + ["ls", "--format", "json"],
             capture_output=True,
             check=True,
         ).stdout
@@ -88,7 +79,7 @@ def compose_stop_project(project_name: str):
     log.info("Stopping running projects")
     for p in compose_list_projects():
         if p["Name"] == project_name:
-            subprocess.run(_docker_compose_bin() + ["-f", p["ConfigFiles"], "down"])
+            subprocess.run(_docker_compose_bin + ["-f", p["ConfigFiles"], "down"])
 
 
 def make_bundle(app_interface_path: Path, qontract_server_path: Path):